Move load_from_run_path and forward_model_ok into LocalEnsemble
