+++ src/ert/libres_facade.py
@@ -1,10 +1,7 @@
 from __future__ import annotations
 
-import asyncio
-import logging
 import warnings
 from collections.abc import Callable, Iterable
-from multiprocessing.pool import ThreadPool
 from pathlib import Path
 from typing import (
     TYPE_CHECKING,
@@ -15,7 +12,6 @@ import numpy as np
 from pandas import DataFrame
 
 from ert.analysis import AnalysisEvent, SmootherSnapshot, smoother_update
-from ert.callbacks import forward_model_ok
 from ert.config import (
     ErtConfig,
     Field,
@@ -23,26 +19,14 @@ from ert.config import (
 )
 from ert.data import MeasuredData
 from ert.data._measured_data import ObservationError, ResponseError
-from ert.load_status import LoadResult, LoadStatus
 
 from .plugins import ErtPluginContext
-
-_logger = logging.getLogger(__name__)
 
 if TYPE_CHECKING:
     from ert.config import (
         EnkfObs,
     )
     from ert.storage import Ensemble, Storage
-
-
-def _load_realization_from_run_path(
-    run_path: str,
-    realization: int,
-    ensemble: Ensemble,
-) -> tuple[LoadResult, int]:
-    result = asyncio.run(forward_model_ok(run_path, realization, 0, ensemble))
-    return result, realization
 
 
 class LibresFacade:
@@ -107,41 +91,6 @@ class LibresFacade:
     @property
     def resolved_run_path(self) -> str:
         return str(Path(self.config.runpath_config.runpath_format_string).resolve())
-
-    @staticmethod
-    def load_from_run_path(
-        run_path_format: str,
-        ensemble: Ensemble,
-        active_realizations: list[int],
-    ) -> int:
-        """Returns the number of loaded realizations"""
-        pool = ThreadPool(processes=8)
-
-        async_result = [
-            pool.apply_async(
-                _load_realization_from_run_path,
-                (
-                    run_path_format.replace("<IENS>", str(realization)).replace(
-                        "<ITER>", "0"
-                    ),
-                    realization,
-                    ensemble,
-                ),
-            )
-            for realization in active_realizations
-        ]
-
-        loaded = 0
-        for t in async_result:
-            ((status, message), iens) = t.get()
-
-            if status == LoadStatus.LOAD_SUCCESSFUL:
-                loaded += 1
-            else:
-                _logger.error(f"Realization: {iens}, load failure: {message}")
-
-        ensemble.refresh_ensemble_state()
-        return loaded
 
     def get_observations(self) -> EnkfObs:
         return self.config.enkf_obs
